requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.ini
#
asgiref==3.5.0            # via django
aspy.yaml==1.3.0          # via pre-commit
asttokens==2.0.5          # via stack-data
attrs==21.4.0             # via flake8-bugbear, flake8-mypy, pytest
backcall==0.2.0           # via ipython
backports.zoneinfo==0.2.1  # via django
black==22.3.0
certifi==2021.10.8        # via requests
cfgv==3.3.1               # via pre-commit
chardet==4.0.0            # via requests
click==8.1.2              # via black, pip-tools
coverage==5.0.4
decorator==5.1.1          # via ipython
distlib==0.3.4            # via virtualenv
dj-database-url==0.5.0
dj-email-url==1.0.5
django-configurations==2.2
django-cors-headers==3.11.0
django-debug-toolbar==2.2.1
django-extensions==2.2.9
django-pyodbc==1.1.3
django==4.0.4
djangorestframework==3.13.1
executing==0.8.3          # via stack-data
filelock==3.6.0           # via tox, virtualenv
flake8-bugbear==20.1.4
flake8-docstrings==1.5.0
flake8-isort==3.0.1
flake8-mypy==17.8.0
flake8==3.8.4
identify==2.5.0           # via pre-commit
idna==2.10                # via requests
iniconfig==1.1.1          # via pytest
invoke==1.4.1
ipdb==0.12.3
ipython==8.2.0            # via ipdb
isort[pyproject]==4.3.21
jedi==0.18.1              # via ipython
matplotlib-inline==0.1.3  # via ipython
mccabe==0.6.1             # via flake8
mypy-extensions==0.4.3    # via black, mypy
mypy==0.942
nodeenv==1.6.0            # via pre-commit
packaging==21.3           # via pytest, tox
parso==0.8.3              # via jedi
pathspec==0.9.0           # via black
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==4.4.1
platformdirs==2.5.2       # via black, virtualenv
pluggy==0.13.1            # via pytest, tox
pre-commit==1.21.0
prompt-toolkit==3.0.29    # via ipython
psycopg2==2.8.6
ptyprocess==0.7.0         # via pexpect
pure-eval==0.2.2          # via stack-data
py==1.11.0                # via pytest, tox
pycodestyle==2.6.0        # via flake8
pydocstyle==6.1.1         # via flake8-docstrings
pydot==1.4.2
pyflakes==2.2.0           # via flake8
pygments==2.12.0          # via ipython
pyodbc==4.0.32
pyparsing==3.0.8          # via packaging, pydot
pytest-django==3.8.0
pytest==7.1.2             # via pytest-django
python-dateutil==2.8.2
python-dotenv==0.10.5
pytz==2022.1              # via djangorestframework
pyyaml==6.0
requests==2.25.1
rope==0.16.0
six==1.16.0               # via asttokens, django-configurations, django-extensions, pip-tools, pre-commit, python-dateutil, tox, virtualenv
snowballstemmer==2.2.0    # via pydocstyle
sqlparse==0.4.2           # via django, django-debug-toolbar
stack-data==0.2.0         # via ipython
testfixtures==6.18.5      # via flake8-isort
toml==0.10.2              # via isort, pre-commit, tox
tomli==2.0.1              # via black, mypy, pytest
tox==3.14.6
traitlets==5.1.1          # via ipython, matplotlib-inline
types-pytz==2021.3.6
typing-extensions==4.2.0  # via black, mypy
urllib3==1.26.9           # via requests
uwsgi==2.0.18
virtualenv==20.14.1       # via pre-commit, tox
wcwidth==0.2.5            # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
